abstract ArrayList<Event> as a separate EventList class
