+++ src/main/java/seedu/commands/ShowUpcomingCommand.java
@@ -1,6 +1,7 @@
 package seedu.commands;
 
 import seedu.events.Event;
+import seedu.events.EventList;
 import seedu.exception.EscException;
 import seedu.subjects.SubjectList;
 
@@ -33,7 +34,8 @@ public class ShowUpcomingCommand extends Command {
      */
     @Override
     public void execute(SubjectList subjectList) {
-        subjectList.listUpcoming(dateRange);
+        EventList events = subjectList.getEventList();
+        events.listUpcoming(dateRange);
     }
 
 }
