+++ src/main/java/seedu/subjects/SubjectList.java
@@ -1,23 +1,20 @@
 package seedu.subjects;
 
-import seedu.events.Event;
 import seedu.duke.UI;
+import seedu.events.EventList;
 import seedu.exception.EscException;
 
-import java.time.LocalDate;
-import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+
 
 public class SubjectList {
     private ArrayList<Subject> subjects;
-    private ArrayList<Event> events;
+    private EventList events;
     UI ui = new UI();
 
     public SubjectList() {
         this.subjects = new ArrayList<Subject>();
-        this.events = new ArrayList<Event>();
+        this.events = new EventList();
     }
 
     /**
@@ -26,19 +23,19 @@ public class SubjectList {
      */
     public SubjectList(ArrayList returnObj) {
         this.subjects = (ArrayList<Subject>) returnObj.get(0);
-        this.events = (ArrayList<Event>) returnObj.get(1);
+        this.events = (EventList) returnObj.get(1);
     }
 
-    public SubjectList(ArrayList<Subject> subjects, ArrayList<Event> eventDates) {
+    public SubjectList(ArrayList<Subject> subjects, EventList events) {
         this.subjects = subjects;
-        this.events = eventDates;
+        this.events = events;
     }
 
     public ArrayList<Subject> getSubjects() {
         return this.subjects;
     }
 
-    public ArrayList<Event> getEvents() {
+    public EventList getEventList() {
         return this.events;
     }
 
@@ -135,82 +132,6 @@ public class SubjectList {
                 System.out.println(j + ". " + subjects.get(i).getSubject());
             }
         }
-    }
-
-    /**
-     * Adds an event to the deck.
-     * @param event Event to be added.
-     */
-    public void addEvent(Event event) throws EscException {
-        events.add(event);
-        System.out.println(event.getTopic() + " has been added.");
-        listUpcoming();
-    }
-
-    /**
-     * Removes an event from the deck.
-     * @param index Index of event to be removed.
-     * @throws EscException if list of events is empty or event item at index does not exit.
-     */
-    public void removeEvent(int index) throws EscException {
-        if (this.events.size() == 0) {
-            throw new EscException("The event list is empty.");
-        }
-        try {
-            System.out.println(events.get(index).getTopic() + " has been removed.");
-            events.remove(index);
-            listUpcoming();
-        } catch (IndexOutOfBoundsException e) {
-            throw new EscException("The event item does not exist.");
-        }
-    }
-
-    /**
-     * Lists all upcoming events.
-     * @param dateRange Date range of upcoming events to show.
-     */
-    public void listUpcoming(int... dateRange) {
-        if (this.events.size() == 0) {
-            System.out.println("No stored events.");
-        }
-        ArrayList<Event> listToShow;
-        if (dateRange.length == 0) {
-            System.out.println("Here is the list of events.");
-            listToShow = this.events;
-        } else {
-            System.out.println("Here is the list of upcoming events within " + dateRange[0] + " day(s).");
-            listToShow = findUpcoming(dateRange[0]);
-        }
-
-        if (listToShow.isEmpty()) {
-            System.out.println("No upcoming event within " + dateRange[0] + " day(s).");
-        } else {
-            for (int i = 0; i < listToShow.size(); i++) {
-                int j = i + 1;
-                System.out.println(j + ". " + listToShow.get(i).toString());
-            }
-        }
-
-    }
-
-    /**
-     * Finds upcoming events within the specified date range.
-     * @param dateRange Date range to find upcoming events.
-     * @return Upcoming events within the date range.
-     */
-    public ArrayList<Event> findUpcoming(int dateRange) {
-        ArrayList<Event> upcomingEvents = new ArrayList<>();
-        LocalDate today = LocalDate.now();
-        for (Event event : events) {
-            LocalDate date = event.getDate();
-            long period = today.until(date, ChronoUnit.DAYS);
-            if (period <= dateRange && period >= 0) {
-                upcomingEvents.add(event);
-            }
-        }
-
-        Collections.sort(upcomingEvents);
-        return upcomingEvents;
     }
 
     public int size() {
